refactor: Move Comparator and Sorting to new Sorting.ts
To help break cyclic dependencies.
