+++ src/Query/Query.ts
@@ -2,7 +2,8 @@ import { LayoutOptions } from '../LayoutOptions';
 import type { Task } from '../Task';
 import type { IQuery } from '../IQuery';
 import { getSettings } from '../Config/Settings';
-import { Sort, Sorting } from './Sort';
+import { Sort } from './Sort';
+import type { Sorting } from './Sorting';
 import type { TaskGroups } from './TaskGroups';
 import { parseFilter, parseSorter } from './FilterParser';
 import { Group } from './Group';
